+++ projects/angular-cesium/src/lib/angular-cesium-widgets/components/hippodrome-editor/hippodrome-editor.component.ts
@@ -17,24 +17,24 @@ import { EditableHippodrome } from '../../models/editable-hippodrome';
 @Component({
   selector: 'hippodrome-editor',
   template: /*html*/ `
-    <ac-layer #editHippodromesLayer acFor="let hippodrome of editHippodromes$" [context]="this">
-      <ac-corridor-desc
-        props="{
-		positions: hippodrome.getRealPositionsCallbackProperty(),
-		cornerType: Cesium.CornerType.ROUNDED,
-		material: hippodrome.hippodromeProps.material,
-		width : hippodrome.hippodromeProps.width,
-		outline: hippodrome.hippodromeProps.outline,
-		outlineColor: hippodrome.hippodromeProps.outlineColor,
-        outlineWidth: hippodrome.hippodromeProps.outlineWidth,
-        height: 0
-	}"
-      >
-      </ac-corridor-desc>
+      <ac-layer #editHippodromesLayer acFor="let hippodrome of editHippodromes$" [context]="this">
+          <ac-corridor-desc props="{
+            positions: hippodrome.getRealPositionsCallbackProperty(),
+            cornerType: Cesium.CornerType.ROUNDED,
+            material: hippodrome.hippodromeProps.material,
+            width : hippodrome.hippodromeProps.width,
+            outline: hippodrome.hippodromeProps.outline,
+            outlineColor: hippodrome.hippodromeProps.outlineColor,
+            outlineWidth: hippodrome.hippodromeProps.outlineWidth,
+            height: 0,
+            classificationType: hippodrome.hippodromeProps.classificationType,
+            zIndex: hippodrome.hippodromeProps.zIndex,
+            shadows: hippodrome.hippodromeProps.shadows,
+    }">
+          </ac-corridor-desc>
 
-      <ac-array-desc acFor="let label of hippodrome.labels" [idGetter]="getLabelId">
-        <ac-label-primitive-desc
-          props="{
+          <ac-array-desc acFor="let label of hippodrome.labels" [idGetter]="getLabelId">
+              <ac-label-primitive-desc props="{
             position: label.position,
             backgroundColor: label.backgroundColor,
             backgroundPadding: label.backgroundPadding,
@@ -55,26 +55,25 @@ import { EditableHippodrome } from '../../models/editable-hippodrome';
             style: label.style,
             text: label.text,
             translucencyByDistance: label.translucencyByDistance,
-            verticalOrigin: label.verticalOrigin
-        }"
-        >
-        </ac-label-primitive-desc>
-      </ac-array-desc>
-    </ac-layer>
+            verticalOrigin: label.verticalOrigin,
+            disableDepthTestDistance: label.disableDepthTestDistance,
+        }">
+              </ac-label-primitive-desc>
+          </ac-array-desc>
+      </ac-layer>
 
-    <ac-layer #editPointsLayer acFor="let point of editPoints$" [context]="this">
-      <ac-point-desc
-        props="{
-        position: point.getPosition(),
-        pixelSize: getPointSize(point),
-        color: point.props.color,
-        outlineColor: point.props.outlineColor,
-        outlineWidth: point.props.outlineWidth,
-        show: getPointShow(point)
-    }"
-      >
-      </ac-point-desc>
-    </ac-layer>
+      <ac-layer #editPointsLayer acFor="let point of editPoints$" [context]="this">
+          <ac-point-desc props="{
+         position: point.getPositionCallbackProperty(),
+         pixelSize: getPointSize(point),
+         color: point.props.color,
+         outlineColor: point.props.outlineColor,
+         outlineWidth: point.props.outlineWidth,
+         show: getPointShow(point),
+         disableDepthTestDistance: point.props.disableDepthTestDistance,
+    }">
+          </ac-point-desc>
+      </ac-layer>
   `,
   providers: [CoordinateConverter, HippodromeManagerService],
   changeDetection: ChangeDetectionStrategy.OnPush,