+++ projects/angular-cesium/src/lib/angular-cesium-widgets/models/editable-ellipse.ts
@@ -31,8 +31,8 @@ export class EditableEllipse extends AcEntity {
     private options: EllipseEditOptions,
   ) {
     super();
-    this._ellipseProps = options.ellipseProps;
-    this._pointProps = options.pointProps;
+    this._ellipseProps = {...options.ellipseProps};
+    this._pointProps = {...options.pointProps};
   }
 
   get labels(): LabelProps[] {
@@ -250,7 +250,8 @@ export class EditableEllipse extends AcEntity {
     const majorRadius = this.getMajorRadius();
     const rotation = this.getRotation();
     const delta = GeoUtilsService.getPositionsDelta(this.lastDraggedToPosition, dragEndPosition);
-    GeoUtilsService.addDeltaToPosition(this.getCenter(), delta, true);
+    const newCenterPosition = GeoUtilsService.addDeltaToPosition(this.getCenter(), delta, true);
+    this._center.setPosition(newCenterPosition);
     this.majorRadiusPoint.setPosition(GeoUtilsService.pointByLocationDistanceAndAzimuth(this.getCenter(), majorRadius, rotation));
     this.updatePointsLayer();
     this.updateMinorRadiusEditPoints();