+++ projects/angular-cesium/src/lib/angular-cesium-widgets/models/polyline-edit-options.ts
@@ -10,11 +10,16 @@ export interface PointProps {
   pixelSize?: number;
   showVirtual?: boolean;
   show?: boolean;
+  disableDepthTestDistance?: number;
 }
 
 export interface PolylineProps {
   width?: number;
   material?: Function;
+
+  clampToGround?: boolean;
+  zIndex?: number;
+  classificationType?: any;
 }
 
 export interface PolylineEditOptions extends EditorEditOptions {
@@ -23,4 +28,5 @@ export interface PolylineEditOptions extends EditorEditOptions {
   removePointEvent?: CesiumEvent;
   removePointModifier?: CesiumEventModifier;
   maximumNumberOfPoints?: number;
+  clampHeightTo3D?: boolean;
 }