+++ projects/angular-cesium/src/lib/angular-cesium/services/viewer-factory/viewer-factory.service.ts
@@ -28,12 +28,6 @@ export class ViewerFactory {
     } else {
       viewer = new this.cesium.Viewer(mapContainer,
         {
-          // Poor internet connection - use default globe image, TODO: should be removed
-          imageryProvider: Cesium.createTileMapServiceImageryProvider({
-            url: Cesium.buildModuleUrl('Assets/Textures/NaturalEarthII')
-          }),
-          baseLayerPicker: false,
-          geocoder: false,
           contextOptions: {
             webgl: {preserveDrawingBuffer: true}
           },