+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 ## 0.57
-### Fixes
+### Features
+* Added `debug` prop to `ac-layer` component for printing received AcNotifications
+* Polygon Editor and Polyline Editor support 3D tiles and terrain, to enable set `clampHeightTo3D` parameter to `true`.
+* Added examples for 3D mode in `polyline-editor-layer.ts` and `polygon-editor-layer.ts`.
+* Add Angular CLI schematics
+
+### Fixes 
+* Fix `CameraService` zoomOut().
+* Fix Shape editors not cloning geometries props 
+* Change Shape editor points to be shown above the lines (uses `disableDepthTestDistance` prop)
 * Fixed demo app unit tests
 * Updated README with instructions for including Cesium scripts in test config
 * Added use of the 'availability' field inside of the Cesium entity API
-* Bumped library version
+
+
+### Breaking Changes
+* Remove default `viewer` configuration (the defualt was setting offline map as the initial imagery).
+* Change editors initial props - different color and outline.
+* Change editors initial props - to use Cesium `CallbackProperties`.
 
 ## 0.56
 ### Features